+++ src/Classifiers/ClassificationTree.php
@@ -227,7 +227,7 @@ class ClassificationTree extends CART implements Estimator, Learner, Probabilist
      */
     protected function terminate(Labeled $dataset) : Best
     {
-        $n = $dataset->numRows();
+        $n = $dataset->numSamples();
 
         $counts = array_count_values($dataset->labels());