+++ src/Classifiers/KDNeighbors.php
@@ -172,7 +172,7 @@ class KDNeighbors implements Estimator, Learner, Probabilistic, Persistable
 
         $this->classes = array_fill_keys($dataset->possibleOutcomes(), 0.0);
 
-        $this->featureCount = $dataset->numColumns();
+        $this->featureCount = $dataset->numFeatures();
 
         $this->tree->grow($dataset);
     }