+++ tests/Datasets/LabeledTest.php
@@ -102,8 +102,8 @@ class LabeledTest extends TestCase
 
         $this->assertInstanceOf(Labeled::class, $dataset);
 
-        $this->assertEquals(3, $dataset->numRows());
-        $this->assertEquals(1, $dataset->numColumns());
+        $this->assertEquals(3, $dataset->numSamples());
+        $this->assertEquals(1, $dataset->numFeatures());
     }
 
     /**
@@ -126,9 +126,9 @@ class LabeledTest extends TestCase
     /**
      * @test
      */
-    public function numRows() : void
+    public function numSamples() : void
     {
-        $this->assertEquals(6, $this->dataset->numRows());
+        $this->assertEquals(6, $this->dataset->numSamples());
     }
 
     /**
@@ -144,9 +144,9 @@ class LabeledTest extends TestCase
     /**
      * @test
      */
-    public function numColumns() : void
+    public function numFeatures() : void
     {
-        $this->assertEquals(4, $this->dataset->numColumns());
+        $this->assertEquals(4, $this->dataset->numFeatures());
     }
 
     /**
@@ -592,7 +592,7 @@ class LabeledTest extends TestCase
      */
     public function join() : void
     {
-        $this->assertEquals(count(current(self::SAMPLES)), $this->dataset->numColumns());
+        $this->assertEquals(count(current(self::SAMPLES)), $this->dataset->numFeatures());
 
         $dataset = new Unlabeled([
             [1],
@@ -605,7 +605,7 @@ class LabeledTest extends TestCase
 
         $joined = $this->dataset->join($dataset);
 
-        $this->assertEquals(count(current(self::SAMPLES)) + 1, $joined->numColumns());
+        $this->assertEquals(count(current(self::SAMPLES)) + 1, $joined->numFeatures());
 
         $this->assertEquals(['mean', 'furry', 'loner', -1.5, 2], $joined->sample(1));
         $this->assertEquals(['nice', 'rough', 'friendly', 2.6, 3], $joined->sample(2));